Update .bash_aliases
'dirs' has been aliased to 'dirs -v', which prints the directory stack
with one entry per line prefixed with its position in the stack.

The git-dir for the 'dotgit' alias has been changed from 'Dotfiles' to
'dotfiles.git', in keeping with the convention of having bare
repository directory names end with the suffix .git.