Add .Xmodmap file
xmodmap is a utility for modifying keymaps and pointer button mappings
in Xorg. The '.Xmodmap' contains the key map table.

In my case, '.Xmodmap' is used to swap Super_L and Caps_Lock, and make
the Menu key also function as Caps_Lock.

The use of this configuration file may be deprecated if I get end up
buying a good mechanical keyboard, in which case I will do most mappings
via the firmware. :^)